requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
openpyxl==3.1.5
python-dotenv==1.0.0
